Updating: adding catalogTileView skeleton
