+++ frontend/public/components/marketplace/index.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef, no-unused-vars */
 
-import React from 'react';
+import * as React from 'react';
 import * as _ from 'lodash-es';
 
- export { KubernetesMarketplace } from './kubernetes-marketplace';
+export { KubernetesMarketplace } from './kubernetes-marketplace';
