+++ frontend/public/style.scss
@@ -7,6 +7,8 @@
 // External dependencies that are required/extended by our custom styles (tilde tells Webpack not to use relative path)
 // Patternfly variables must be imported before Bootstrap
 @import '~patternfly/dist/sass/patternfly/variables';
+@import "~patternfly/dist/sass/patternfly/dependencies/bootstrap-switch/variables";
+@import '~patternfly-react-extensions/dist/sass/patternfly-react-extensions';
 @import "~bootstrap-sass/assets/stylesheets/bootstrap/variables";
 @import "~bootstrap-sass/assets/stylesheets/bootstrap/mixins";
 @import "~font-awesome-sass/assets/stylesheets/font-awesome/variables";
